requirements.txt
APScheduler == 3.6.3
plurk_oauth == 0.9.2
python_dotenv == 0.21.0
pytz == 2021.3
requests == 2.22.0
